# SILO-plateforme-source-2026-07-27/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 3C Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List transactions (role-shaped amounts)
      description: Admin/PM see full split; partners only their 70% share; clients only their totals
      responses:
        "200":
          description: Transactions visible to the caller
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Prepare a payment request (admin/PM) — split computed server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Pending payment created with server-computed split
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error (e.g. below quote floor)
        "403":
          description: Forbidden
  /transactions/{id}/checkout:
    post:
      operationId: createTransactionCheckout
      tags: [finance]
      summary: Create or resume the authenticated client's Stripe Checkout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stripe-hosted payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionCheckout"
        "403":
          description: Forbidden
        "404":
          description: Transaction not found
        "409":
          description: Transaction is no longer payable
        "502":
          description: Payment provider unavailable
        "503":
          description: Payment provider not configured
  /frp/accounts:
    get:
      operationId: listFrpAccounts
      tags: [finance]
      summary: List FRP accounts (admin/PM all, partner own only)
      responses:
        "200":
          description: FRP accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FrpAccount"
        "403":
          description: Forbidden
  /frp/movements:
    get:
      operationId: listFrpMovements
      tags: [finance]
      summary: List FRP movements (admin/PM all, partner own only)
      responses:
        "200":
          description: FRP movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FrpMovement"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Transaction:
      type: object
      required:
        [
          id,
          title,
          agencyName,
          kind,
          status,
          calculationVersion,
          date,
          paidAt,
          createdAt,
        ]
      properties:
        id:
          type: integer
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        clientUserId:
          type: ["string", "null"]
        agencyId:
          type: ["string", "null"]
        agencyName:
          type: string
        advisorUserId:
          type: ["string", "null"]
        kind:
          type: string
          enum: [ponctuel, abonnement]
        status:
          type: string
          enum: [en_attente_paiement, confirmee, echouee, annulee, remboursee]
        amountHt:
          type: ["number", "null"]
        tva:
          type: ["number", "null"]
        ttc:
          type: ["number", "null"]
        partAgence:
          type: ["number", "null"]
        partSilo:
          type: ["number", "null"]
        partFrp:
          type: ["number", "null"]
        fraisPspTotal:
          type: ["number", "null"]
        fraisPspAgence:
          type: ["number", "null"]
        fraisPspSilo:
          type: ["number", "null"]
        partAgenceApresPsp:
          type: ["number", "null"]
        partSiloApresPsp:
          type: ["number", "null"]
        reserveIncidents:
          type: ["number", "null"]
        primeConseiller:
          type: ["number", "null"]
        contributionSiloApresVariables:
          type: ["number", "null"]
        calculationVersion:
          type: string
        date:
          type: string
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TransactionInput:
      type: object
      required: [projectId, title, agencyId, kind, amountHt]
      properties:
        projectId:
          type: integer
        title:
          type: string
          minLength: 1
        agencyId:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [ponctuel, abonnement]
        amountHt:
          type: number
          exclusiveMinimum: 0
        date:
          type: ["string", "null"]
    TransactionCheckout:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
    FrpAccount:
      type: object
      required: [agencyId, agencyName, balance, transactionsThisYear, outcome]
      properties:
        agencyId:
          type: string
        agencyName:
          type: string
        balance:
          type: number
        transactionsThisYear:
          type: integer
        outcome:
          type: string
          enum: [decision_manuelle_requise]
    FrpMovement:
      type: object
      required: [id, agencyId, agencyName, date, label, type, amount]
      properties:
        id:
          type: integer
        transactionId:
          type: ["integer", "null"]
        agencyId:
          type: string
        agencyName:
          type: string
        date:
          type: string
        label:
          type: string
        projectTitle:
          type: ["string", "null"]
        type:
          type: string
          enum: [credit, reversement, reinvestissement]
        amount:
          type: number
...